requirements.txt
requests
python-dotenv
colorama
pywin32 ; platform_system == "Windows"
